Extract CLI logic to Engine::Game instead
This helps keep the cli.rb file small and additions there are focused on
the interface messages instead of game functionality.

Removed the Engine::Pile in favour of more logic within the
Engine::Game.

Introduced the Engine::Play which checks for the validity of plays of
users and between users.