Add round-end calculations
Extracted each round logic into Engine::Round objects, with a
Engine::Current object that does state tracking for certain append-only
variables, such as the pile, and the plays.

The Engine::Game also now handles correctly the situation at the end of
a round, dealing with initiative and the current priority card.